requirements.txt
pandas
numpy<2.0.0
tensorflow
scikit-learn
keras-preprocessing
datasets